+++ crates/ton/src/tep.rs
@@ -1,7 +1,0 @@
-pub mod excesses_msg;
-pub mod jetton;
-pub mod metadata;
-pub mod mintless_airdrop;
-pub mod nft;
-pub mod sbt;
-pub mod snake_data;