+++ crates/ton/src/tep/metadata.rs
@@ -1,9 +1,0 @@
-mod meta_loader;
-mod metadata_content;
-mod metadata_fields;
-mod traits;
-
-pub use meta_loader::*;
-pub use metadata_content::*;
-pub use metadata_fields::*;
-pub use traits::*;